+++ front-end-app/app/src/main/java/com/dev/hirelink/dto/BasicErrorResponse.kt
@@ -1,0 +1,8 @@
+package com.dev.hirelink.dto
+
+data class BasicErrorResponse(
+    val title: String? = null,
+    val description: String? = null,
+    val violations: List<ConstraintViolation>? = null,
+    val message: String? = null
+)
