+++ front-end-app/app/src/main/java/com/dev/hirelink/dto/ConstraintViolation.kt
@@ -1,0 +1,7 @@
+package com.dev.hirelink.dto
+
+data class ConstraintViolation(
+    val propertyPath: String? = null,
+    val message: String? = null,
+    val code: String? = null
+)
