+++ front-end-app/app/src/main/java/com/dev/hirelink/dto/EmployerChoosePlanRequest.kt
@@ -1,3 +1,3 @@
-package com.dev.hirelink.modules.auth.dto
+package com.dev.hirelink.dto
 
 data class EmployerChoosePlanRequest(val plan: String)
