+++ front-end-app/app/src/main/java/com/dev/hirelink/modules/auth/register/fragments/employerregister/EmployerRegisterStep1Fragment.kt
@@ -15,6 +15,8 @@ import com.dev.hirelink.R
 import com.dev.hirelink.components.HttpExceptionParser
 import com.dev.hirelink.components.SharedPreferenceManager
 import com.dev.hirelink.databinding.FragmentEmployerRegisterStep1Binding
+import com.dev.hirelink.dto.ApplicationUserRequest
+import com.dev.hirelink.dto.BasicErrorResponse
 import com.dev.hirelink.enums.RegistrationStep
 import com.dev.hirelink.enums.RoleType
 import com.dev.hirelink.models.*
