+++ front-end-app/app/src/main/java/com/dev/hirelink/modules/auth/register/fragments/employerregister/EmployerRegisterStep2Fragment.kt
@@ -9,21 +9,19 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.core.os.bundleOf
 import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
 import com.dev.hirelink.R
 import com.dev.hirelink.components.HttpExceptionParser
 import com.dev.hirelink.components.SharedPreferenceManager
 import com.dev.hirelink.databinding.FragmentEmployerRegisterStep2Binding
 import com.dev.hirelink.enums.RegistrationStep
-import com.dev.hirelink.models.BasicErrorResponse
-import com.dev.hirelink.modules.auth.dto.OTPVerificationRequest
+import com.dev.hirelink.dto.BasicErrorResponse
+import com.dev.hirelink.dto.OTPVerificationRequest
 import com.dev.hirelink.modules.auth.register.RegisterActivity
 import com.dev.hirelink.modules.auth.register.fragments.StepFragment
 import com.dev.hirelink.modules.auth.register.fragments.candidateregister.RegisterViewModel
 import com.dev.hirelink.modules.auth.register.fragments.employerregister.EmployerRegisterStep1Fragment.Companion.NATIONAL_UNIQUE_NUMBER_KEY
 import com.dev.hirelink.modules.auth.register.fragments.employerregister.EmployerRegisterStep1Fragment.Companion.USER_EMAIL_KEY
 import com.dev.hirelink.modules.auth.register.fragments.employerregister.EmployerRegisterStep1Fragment.Companion.USER_ID_KEY
-import com.dev.hirelink.modules.auth.viewmodel.AuthViewModel
 import com.dev.hirelink.modules.common.CustomLoadingOverlay
 import com.google.android.material.snackbar.Snackbar
 import io.reactivex.android.schedulers.AndroidSchedulers
