+++ front-end-app/app/src/main/java/com/dev/hirelink/modules/auth/register/fragments/employerregister/EmployerRegisterStep4Fragment.kt
@@ -2,32 +2,25 @@ package com.dev.hirelink.modules.auth.register.fragments.employerregister
 
 import android.os.Bundle
 import android.util.Log
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.os.bundleOf
 import androidx.databinding.DataBindingUtil
-import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.GridLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.dev.hirelink.R
 import com.dev.hirelink.components.HttpExceptionParser
 import com.dev.hirelink.components.SharedPreferenceManager
-import com.dev.hirelink.data.MockDataSource
 import com.dev.hirelink.databinding.FragmentEmployerRegisterStep4Binding
 import com.dev.hirelink.enums.RegistrationStep
 import com.dev.hirelink.models.ApplicationUser
-import com.dev.hirelink.models.BasicErrorResponse
+import com.dev.hirelink.dto.BasicErrorResponse
 import com.dev.hirelink.models.Plan
-import com.dev.hirelink.modules.auth.dto.EmployerChoosePlanRequest
+import com.dev.hirelink.dto.EmployerChoosePlanRequest
 import com.dev.hirelink.modules.auth.register.RegisterActivity
 import com.dev.hirelink.modules.auth.register.adapters.PlanItemAdapter
 import com.dev.hirelink.modules.auth.register.fragments.StepFragment
 import com.dev.hirelink.modules.auth.register.fragments.candidateregister.RegisterViewModel
-import com.dev.hirelink.modules.auth.register.fragments.employerregister.EmployerRegisterStep1Fragment.Companion.NATIONAL_UNIQUE_NUMBER_KEY
-import com.dev.hirelink.modules.auth.register.fragments.employerregister.EmployerRegisterStep1Fragment.Companion.USER_EMAIL_KEY
-import com.dev.hirelink.modules.auth.register.fragments.employerregister.EmployerRegisterStep1Fragment.Companion.USER_ID_KEY
 import com.dev.hirelink.modules.common.CustomLoadingOverlay
 import com.google.android.material.snackbar.Snackbar
 import io.reactivex.android.schedulers.AndroidSchedulers
