+++ front-end-app/app/src/main/java/com/dev/hirelink/modules/auth/register/fragments/employerregister/EmployerRegisterStep6DirectDebitFragment.kt
@@ -4,28 +4,23 @@ import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
 import android.util.Log
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.os.bundleOf
 import androidx.databinding.DataBindingUtil
 import com.dev.hirelink.R
 import com.dev.hirelink.components.HttpExceptionParser
 import com.dev.hirelink.components.SharedPreferenceManager
 import com.dev.hirelink.databinding.FragmentEmployerRegisterStep6DirectDebitBinding
-import com.dev.hirelink.enums.PaymentType
 import com.dev.hirelink.enums.RegistrationStep
 import com.dev.hirelink.models.ApplicationUser
-import com.dev.hirelink.models.BasicErrorResponse
-import com.dev.hirelink.models.PaymentInformationDTO
-import com.dev.hirelink.modules.auth.dto.BankInformationDTO
+import com.dev.hirelink.dto.BasicErrorResponse
+import com.dev.hirelink.dto.BankInformationDTO
 import com.dev.hirelink.modules.auth.register.RegisterActivity
 import com.dev.hirelink.modules.auth.register.fragments.StepFragment
 import com.dev.hirelink.modules.auth.register.fragments.candidateregister.RegisterViewModel
 import com.dev.hirelink.modules.common.CustomLoadingOverlay
 import com.google.android.material.snackbar.Snackbar
-import com.vicmikhailau.maskededittext.MaskedFormatter
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.schedulers.Schedulers
