+++ front-end-app/app/src/main/java/com/dev/hirelink/modules/auth/viewmodel/AuthViewModel.kt
@@ -2,7 +2,7 @@ package com.dev.hirelink.modules.auth.viewmodel
 
 import androidx.lifecycle.ViewModel
 import com.dev.hirelink.models.ApplicationUser
-import com.dev.hirelink.modules.auth.dto.Credentials
+import com.dev.hirelink.dto.Credentials
 import com.dev.hirelink.network.auth.AuthRepository
 
 class AuthViewModel(private val repository: AuthRepository) : ViewModel() {
