+++ front-end-app/app/src/main/java/com/dev/hirelink/network/auth/AuthService.kt
@@ -1,13 +1,12 @@
 package com.dev.hirelink.network.auth
 
 import com.dev.hirelink.models.ApplicationUser
-import com.dev.hirelink.models.ApplicationUserRequest
-import com.dev.hirelink.modules.auth.dto.Credentials
-import com.dev.hirelink.modules.auth.dto.EmployerCreatePasswordRequest
-import com.dev.hirelink.modules.auth.dto.LoginResponse
-import com.dev.hirelink.modules.auth.dto.OTPVerificationRequest
+import com.dev.hirelink.dto.ApplicationUserRequest
+import com.dev.hirelink.dto.Credentials
+import com.dev.hirelink.dto.EmployerCreatePasswordRequest
+import com.dev.hirelink.dto.LoginResponse
+import com.dev.hirelink.dto.OTPVerificationRequest
 import io.reactivex.Single
-import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.Headers
