+++ front-end-app/app/src/main/java/com/dev/hirelink/network/bankinformation/BankInformationRepository.kt
@@ -3,7 +3,7 @@ package com.dev.hirelink.network.bankinformation
 import android.content.Context
 import com.dev.hirelink.components.RetrofitServiceBuilder
 import com.dev.hirelink.models.BankInformation
-import com.dev.hirelink.modules.auth.dto.BankInformationDTO
+import com.dev.hirelink.dto.BankInformationDTO
 
 class BankInformationRepository(private val context: Context) {
     private val service: BankInformationService by lazy {
