+++ front-end-app/app/src/main/java/com/dev/hirelink/network/bankinformation/BankInformationService.kt
@@ -1,7 +1,7 @@
 package com.dev.hirelink.network.bankinformation
 
 import com.dev.hirelink.models.BankInformation
-import com.dev.hirelink.modules.auth.dto.BankInformationDTO
+import com.dev.hirelink.dto.BankInformationDTO
 import io.reactivex.Single
 import retrofit2.http.Body
 import retrofit2.http.POST
