+++ front-end-app/app/src/main/java/com/dev/hirelink/network/paymentinformation/PaymentInformationRepository.kt
@@ -2,8 +2,7 @@ package com.dev.hirelink.network.paymentinformation
 
 import android.content.Context
 import com.dev.hirelink.components.RetrofitServiceBuilder
-import com.dev.hirelink.models.PaymentInformation
-import com.dev.hirelink.models.PaymentInformationDTO
+import com.dev.hirelink.dto.PaymentInformationDTO
 
 class PaymentInformationRepository(private val context: Context) {
     private val service: PaymentInformationService by lazy {
