+++ front-end-app/app/src/main/java/com/dev/hirelink/network/paymentinformation/PaymentInformationService.kt
@@ -1,7 +1,7 @@
 package com.dev.hirelink.network.paymentinformation
 
 import com.dev.hirelink.models.PaymentInformation
-import com.dev.hirelink.models.PaymentInformationDTO
+import com.dev.hirelink.dto.PaymentInformationDTO
 import io.reactivex.Single
 import retrofit2.http.Body
 import retrofit2.http.POST
