+++ front-end-app/app/src/main/java/com/dev/hirelink/network/plan/PlanRepository.kt
@@ -3,8 +3,6 @@ package com.dev.hirelink.network.plan
 import android.content.Context
 import com.dev.hirelink.components.RetrofitServiceBuilder
 import com.dev.hirelink.models.Plan
-import com.dev.hirelink.models.Role
-import com.dev.hirelink.models.RoleType
 import io.reactivex.Single
 
 class PlanRepository(private val context: Context) {
