+++ front-end-app/app/src/main/java/com/dev/hirelink/network/user/UserService.kt
@@ -1,7 +1,7 @@
 package com.dev.hirelink.network.user
 
 import com.dev.hirelink.models.ApplicationUser
-import com.dev.hirelink.modules.auth.dto.EmployerChoosePlanRequest
+import com.dev.hirelink.dto.EmployerChoosePlanRequest
 import io.reactivex.Single
 import retrofit2.http.Body
 import retrofit2.http.PUT
